+++ apis/python/HISTORY.md
@@ -37,6 +37,8 @@ The primary change in 1.17.0 is the upgrade to TileDB 2.28.
 - \[[#4050](https://github.com/single-cell-data/TileDB-SOMA/pull/4050)\] DataFrame `count` and SparseNDArray `nnz` fix - report correct number of cells in array in the case where a delete query had been previously applied.
 - \[[#4055](https://github.com/single-cell-data/TileDB-SOMA/pull/4055)\] Various `open()` code paths failed to check the SOMA encoding version number, and would fail with cryptic errors.
 
+- \[[#4031](https://github.com/single-cell-data/TileDB-SOMA/pull/4031)\] [python] Encode URI when generating from key
+
 ## [Release prior to 1.17.0]
 
 TileDB-SOMA Python releases prior to 1.17.0 are documented in the [TileDB-SOMA Github Releases](https://github.com/single-cell-data/TileDB-SOMA/releases).